+++ CHANGELOG.md
@@ -62,6 +62,7 @@ This is a major release, and it might be not compatible with your current usage 
 - `autoCompleteFieldUtils`: use `suggestFieldUtils`
 - `CodeMirror.outerDivAttributes`: use all properties directly on `CodeEditor`
 - `MultiSuggestField.ofType`: use `<MultiSuggestField<TYPE>>`
+- `StringPreviewContentBlobToggler.firstNonEmptyLineOnly`: use `useOnly="firstNonEmptyLine"`
 - color configuration is not possible anymore via previously used SCSS variables
     - `$eccgui-color-primary`, `$eccgui-color-primary-contrast`, `$eccgui-color-accent`, `$eccgui-color-accent-contrast`,
       `$eccgui-color-success-text`, `$eccgui-color-success-background`, `$eccgui-color-info-text`, `$eccgui-color-info-background`,